requirements_27_05_2025.txt
absl-py==2.1.0
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blis==1.1.0
bs4==0.0.2
captum==0.6.0
Cartopy==0.24.1
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
cftime==1.6.4.post1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.10
dask==2022.7.1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
docutils==0.21.2
efficientnet_pytorch==0.7.1
executing==2.1.0
fastai==2.7.18
fastcore==1.7.28
fastdownload==0.0.7
fasteners==0.19
fastjsonschema==2.21.1
fastprogress==1.0.3
filelock==3.16.1
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
gpytorch==1.13
grpcio==1.68.1
huggingface-hub==0.27.0
icecream==2.1.3
idna==3.10
imageio==2.36.1
imagesize==1.4.1
imbalanced-learn==0.13.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
jaxtyping==0.2.19
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.0.13
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
linear-operator==0.5.3
llvmlite==0.43.0
locket==1.0.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.1.0
munch==4.0.0
murmurhash==1.0.11
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netCDF4==1.7.2
networkx==3.4.2
notebook==7.0.8
notebook_shim==0.2.4
numba==0.60.0
numcodecs==0.14.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
preshed==3.0.9
pretrainedmodels==0.7.4
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pyDeprecate==0.3.2
Pygments==2.18.0
pyod==2.0.3
pyparsing==3.2.0
pyproj==3.7.0
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-docs-theme==2024.12
python-json-logger==3.2.1
pytorch-lightning==1.7.7
pyts==0.13.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
safetensors==0.4.5
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.15.0rc2
seaborn==0.12.2
segmentation-models-pytorch==0.3.4
Send2Trash==1.8.3
sentry-sdk==2.19.2
setproctitle==1.3.4
setuptools==75.6.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
srsly==2.5.0
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
terminado==0.18.1
thinc==8.3.3
threadpoolctl==3.5.0
tifffile==2024.12.12
timm==0.9.7
tinycss2==1.4.0
toolz==1.0.0
torch==2.2.2
torchmetrics==0.11.1
torchvision==0.17.2
tornado==6.4.2
tqdm==4.64.0
traitlets==5.14.3
triton==3.1.0
tsai==0.3.9
typeguard==4.4.1
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.3.0
wandb==0.19.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.0
xarray==2022.6.0
yarl==1.18.3
zarr==2.12.0
